Made common functions for calculate expenses
