+++ Splito/UI/Home/Groups/Group/Group Options/Settle up/GroupSettleUpViewModel.swift
@@ -61,7 +61,6 @@ class GroupSettleUpViewModel: BaseViewModel, ObservableObject {
                 self.members = members
                 self.members.removeAll(where: { $0.id == user.id })
                 self.fetchTransactions()
-                self.fetchExpenses()
             }.store(in: &cancelable)
     }
 
@@ -73,6 +72,7 @@ class GroupSettleUpViewModel: BaseViewModel, ObservableObject {
         } receiveValue: { [weak self] transactions in
             guard let self else { return }
             self.transactions = transactions
+            self.fetchExpenses()
         }.store(in: &cancelable)
     }
 
@@ -83,124 +83,15 @@ class GroupSettleUpViewModel: BaseViewModel, ObservableObject {
                     self?.handleServiceError(error)
                 }
             } receiveValue: { [weak self] expenses in
-                guard let self, let group else { return }
+                guard let self, let group, let userId = preference.user?.id else { return }
                 self.expenses = expenses
+                memberOwingAmount = [:]
                 if group.isDebtSimplified {
-                    self.calculateExpensesSimply()
+                    self.memberOwingAmount = calculateExpensesSimplify(userId: userId, expenses: expenses, transactions: transactions)
                 } else {
-                    self.calculateExpenses()
+                    self.memberOwingAmount = calculateExpensesNonSimplify(userId: userId, expenses: expenses, transactions: transactions)
                 }
             }.store(in: &cancelable)
-    }
-
-    private func calculateExpenses() {
-        guard let userId = preference.user?.id else { return }
-
-        var owesToUser: [String: Double] = [:]
-        var owedByUser: [String: Double] = [:]
-
-        memberOwingAmount = [:]
-
-        for expense in expenses {
-            let splitAmount = expense.amount / Double(expense.splitTo.count)
-
-            if expense.paidBy == userId {
-                for member in expense.splitTo where member != userId {
-                    owesToUser[member, default: 0.0] += splitAmount
-                }
-            } else if expense.splitTo.contains(userId) {
-                owedByUser[expense.paidBy, default: 0.0] += splitAmount
-            }
-        }
-
-        for transaction in transactions {
-            let payer = transaction.payerId
-            let receiver = transaction.receiverId
-            let amount = transaction.amount
-
-            if transaction.payerId == userId {
-                if owedByUser[receiver] != nil {
-                    owesToUser[transaction.receiverId, default: 0.0] += amount
-                } else {
-                    owedByUser[transaction.payerId, default: 0.0] -= amount
-                }
-            } else if transaction.receiverId == userId {
-                if owesToUser[payer] != nil {
-                    owedByUser[transaction.payerId, default: 0.0] += amount
-                } else {
-                    owesToUser[payer] = -amount
-                }
-            }
-        }
-
-        owesToUser.forEach { payerId, owesAmount in
-            memberOwingAmount[payerId, default: 0.0] += owesAmount
-        }
-        owedByUser.forEach { receiverId, owedAmount in
-            memberOwingAmount[receiverId, default: 0.0] -= owedAmount
-        }
-
-        memberOwingAmount = memberOwingAmount.filter { $0.value != 0 }
-    }
-
-    private func calculateExpensesSimply() {
-        guard let userId = preference.user?.id else { return }
-
-        var ownAmounts: [String: Double] = [:]
-
-        memberOwingAmount = [:]
-
-        for expense in expenses {
-            ownAmounts[expense.paidBy, default: 0.0] += expense.amount
-            let splitAmount = expense.amount / Double(expense.splitTo.count)
-
-            for member in expense.splitTo {
-                ownAmounts[member, default: 0.0] -= splitAmount
-            }
-        }
-
-        let debts = settleDebts(users: ownAmounts)
-        for debt in debts where debt.0 == userId || debt.1 == userId {
-            memberOwingAmount[debt.1 == userId ? debt.0 : debt.1] = debt.1 == userId ? debt.2 : -debt.2
-        }
-
-        for transaction in transactions {
-            let payer = transaction.payerId
-            let receiver = transaction.receiverId
-            let amount = transaction.amount
-
-            if payer == userId, let currentAmount = memberOwingAmount[receiver] {
-                memberOwingAmount[receiver] = currentAmount + amount
-            } else if receiver == userId, let currentAmount = memberOwingAmount[payer] {
-                memberOwingAmount[payer] = currentAmount - amount
-            }
-        }
-
-        memberOwingAmount = memberOwingAmount.filter { $0.value != 0 }
-    }
-
-    private func settleDebts(users: [String: Double]) -> [(String, String, Double)] {
-        var mutableUsers = users
-        var debts: [(String, String, Double)] = []
-        let positiveAmounts = mutableUsers.filter { $0.value > 0 }
-        let negativeAmounts = mutableUsers.filter { $0.value < 0 }
-
-        for (creditor, creditAmount) in positiveAmounts {
-            var remainingCredit = creditAmount
-
-            for (debtor, debtAmount) in negativeAmounts {
-                if remainingCredit == 0 { break }
-                let amountToSettle = min(remainingCredit, -debtAmount)
-                if amountToSettle > 0 {
-                    debts.append((debtor, creditor, amountToSettle))
-                    remainingCredit -= amountToSettle
-                    mutableUsers[debtor]! += amountToSettle
-                    mutableUsers[creditor]! -= amountToSettle
-                }
-            }
-        }
-
-        return debts
     }
 
     func getMemberDataBy(id: String) -> AppUser? {
