+++ Splito/UI/Home/Groups/GroupListViewModel.swift
@@ -162,117 +162,17 @@ class GroupListViewModel: BaseViewModel, ObservableObject {
     private func calculateExpenses(expenses: [Expense], transactions: [Transactions]) -> AnyPublisher<(Double, [String: Double]), ServiceError> {
         guard let userId = self.preference.user?.id else { return Fail(error: .dataNotFound).eraseToAnyPublisher() }
 
-        var expenseByUser = 0.0
-        var owesToUser: [String: Double] = [:]
-        var owedByUser: [String: Double] = [:]
-        var ownAmount: [String: Double] = [:]
+        let memberOwingAmount = calculateExpensesNonSimplify(userId: userId, expenses: expenses, transactions: transactions)
 
-        for expense in expenses {
-            let splitAmount = expense.amount / Double(expense.splitTo.count)
-            if expense.paidBy == userId {
-                expenseByUser += expense.splitTo.contains(userId) ? expense.amount - splitAmount : expense.amount
-                for member in expense.splitTo where member != userId {
-                    owesToUser[member, default: 0.0] += splitAmount
-                }
-            } else if expense.splitTo.contains(where: { $0 == userId }) {
-                expenseByUser -= splitAmount
-                owedByUser[expense.paidBy, default: 0.0] += splitAmount
-            }
-        }
-
-        for transaction in transactions {
-            let payer = transaction.payerId
-            let receiver = transaction.receiverId
-            let amount = transaction.amount
-
-            if transaction.payerId == userId {
-                if owedByUser[receiver] != nil {
-                    owesToUser[transaction.receiverId, default: 0.0] += amount
-                } else {
-                    owedByUser[transaction.payerId, default: 0.0] -= amount
-                }
-            } else if transaction.receiverId == userId {
-                if owesToUser[payer] != nil {
-                    owedByUser[transaction.payerId, default: 0.0] += amount
-                } else {
-                    owesToUser[payer] = -amount
-                }
-            }
-        }
-
-        owesToUser.forEach { userId, owesAmount in
-            ownAmount[userId] = owesAmount
-        }
-        owedByUser.forEach { userId, owedAmount in
-            ownAmount[userId] = (ownAmount[userId] ?? 0) - owedAmount
-        }
-        expenseByUser = ownAmount.values.reduce(0, +)
-
-        return Just((expenseByUser, ownAmount)).setFailureType(to: ServiceError.self).eraseToAnyPublisher()
+        return Just((memberOwingAmount.values.reduce(0, +), memberOwingAmount)).setFailureType(to: ServiceError.self).eraseToAnyPublisher()
     }
 
     private func calculateExpensesSimply(expenses: [Expense], transactions: [Transactions]) -> AnyPublisher<(Double, [String: Double]), ServiceError> {
         guard let userId = self.preference.user?.id else { return Fail(error: .dataNotFound).eraseToAnyPublisher() }
 
-        var expenseByUser = 0.0
-        var ownAmount: [String: Double] = [:]
+        let oweByUser = calculateExpensesSimplify(userId: userId, expenses: expenses, transactions: transactions)
 
-        for expense in expenses {
-            ownAmount[expense.paidBy, default: 0.0] += expense.amount
-            let splitAmount = expense.amount / Double(expense.splitTo.count)
-
-            for member in expense.splitTo {
-                ownAmount[member, default: 0.0] -= splitAmount
-            }
-        }
-
-        var oweByUser: [String: Double] = [:]
-        let debts = settleDebts(users: ownAmount)
-
-        for debt in debts where debt.0 == userId || debt.1 == userId {
-            expenseByUser += debt.1 == userId ? debt.2 : -debt.2
-            oweByUser[debt.1 == userId ? debt.0 : debt.1] = debt.1 == userId ? debt.2 : -debt.2
-        }
-
-        for transaction in transactions {
-            let payer = transaction.payerId
-            let receiver = transaction.receiverId
-            let amount = transaction.amount
-
-            if payer == userId, let currentAmount = oweByUser[receiver] {
-                oweByUser[receiver] = currentAmount + amount
-            } else if receiver == userId, let currentAmount = oweByUser[payer] {
-                oweByUser[payer] = currentAmount - amount
-            }
-        }
-
-        expenseByUser = oweByUser.values.reduce(0, +)
-
-        return Just((expenseByUser, oweByUser)).setFailureType(to: ServiceError.self).eraseToAnyPublisher()
-    }
-
-    private func settleDebts(users: [String: Double]) -> [(String, String, Double)] {
-        var mutableUsers = users
-        var debts: [(String, String, Double)] = []
-        let positiveAmounts = mutableUsers.filter { $0.value > 0 }
-        let negativeAmounts = mutableUsers.filter { $0.value < 0 }
-
-        for (creditor, creditAmount) in positiveAmounts {
-            var remainingCredit = creditAmount
-
-            for (debtor, debtAmount) in negativeAmounts {
-                if remainingCredit == 0 { break }
-                let amountToSettle = min(remainingCredit, -debtAmount)
-                if amountToSettle > 0 {
-                    debts.append((debtor, creditor, amountToSettle))
-                    remainingCredit -= amountToSettle
-                    mutableUsers[debtor]! += amountToSettle
-                    mutableUsers[creditor]! -= amountToSettle
-                }
-            }
-        }
-
-        return debts
+        return Just((oweByUser.values.reduce(0, +), oweByUser)).setFailureType(to: ServiceError.self).eraseToAnyPublisher()
     }
 }
